services/classification_service/requirements.txt
fastapi
uvicorn
httpx
pydantic
tenacity
pandas
openpyxl
python-dotenv
celery[redis]
redis
requests
slowapi
